api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis-py-cluster>=1.3.5",
    "redis[hiredis]>=6.2.0",
    "slowapi>=0.1.9",
    "sqlparse>=0.5.3",
    "supabase>=2.17.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
]
